columnar: restructure as a storage subsystem — data + collection + trace faces
Rework the columnar work into a self-contained storage subsystem rooted at
`columnar/`, organized by the function DD needs rather than by implementation:

  columnar/
    layout · updates · trie_merger   DATA — the shared columnar core
    collection/                      COLLECTION face: RecordedUpdates (+ Negate/
                                     Enter/Leave/ResultsIn), Builder, Pact, and
                                     the columnar operators (join_function, ...)
    trace/                           TRACE face: ColChunk (a `Chunk` impl) with
                                     Spine / Batcher / Builder / Chunker, + spill

Both faces derive from the same data; they are linked only by two bridge
operators (the trace `Chunker`, collection->batch; and `as_recorded_updates`,
trace->collection). The generic `Chunk` trait + harness stay in `trace/chunk/`
(layout-agnostic, shared with `vec`); `columnar::trace::ColChunk` implements it.

This supersedes the interim placement under `trace/chunk/col/`: the collection
surface (RecordedUpdates, join_function, ...) was mis-homed inside the
trace/chunk module when it is collection — not chunk — machinery. Rooting
everything under `columnar/` makes it a storage variant that borrows DD's
abstractions but can be sheared off without changing DD's own structure — a
template for future storage variants.

Exports use BATCH/TRACE vocabulary (Spine/Batcher/Builder/Chunker), dropping the
historical `Val*` prefix; `ColChunk` keeps its name as the honest `Chunk` impl.
Also folds in the wire/spill codec dedup (`RecordedUpdates` delegates its body to
`Updates::write_to`/`read_from`) and its round-trip test.

Examples (`columnar`, `columnar_spill`, `spines`) and `chunk_bench` updated to
the new paths. Pure motion otherwise: build/tests clean, runtime smokes unchanged.
